requirements.txt
flask
openai
flask_cors
python-dotenv
